Remove cJSON symbolic link and build as C++
Previously the library implementation was at `src/deps/cjson/cJSON.c`
and there was a symbolic link to it from `src/deps/cjson/cJSON.cpp`.

This had an odd effect on the build. The `.cpp` file was used to build
the `.d` dependency file, while the `.c` file was actually used to build
the `.o` object file:

    g++ -Isrc -Isrc/deps -MM \
        -MF src/deps/cjson/cJSON.d \
        -MT src/deps/cjson/cJSON.d \
        -MT src/deps/cjson/cJSON.o \
        src/deps/cjson/cJSON.cpp

    ...

    cc -c -o \
       src/deps/cjson/cJSON.o \
       src/deps/cjson/cJSON.c

This invocation of `cc` comes from GNU Make's implicit rules, which
is why it lacks all of our flags including optimisation level and
warning options.

This commit removes the symlink, and renames the `.c` source file
to `.cpp`. The effect of this is that the cJSON source is now built
with all the same flags as the rest of the project:

    g++ -Isrc -Isrc/deps -MM \
        -MF src/deps/cjson/cJSON.d \
        -MT src/deps/cjson/cJSON.d \
        -MT src/deps/cjson/cJSON.o \
        src/deps/cjson/cJSON.cpp

    ...

    g++ -std=c++11 -Wall -Wno-reorder -Wno-class-memaccess \
        -Wno-format-truncation -pedantic -DMONO -DSDL \
        -Isrc -Isrc/deps -O3 -DLINUX -c \
        -o src/deps/cjson/cJSON.o \
        src/deps/cjson/cJSON.cpp

This is more explicit and consistent and should have the effect
of optimising load/save that uses cJSON.
